support auto expired py worker
