+++ be/src/common/config.h
@@ -1314,9 +1314,14 @@ CONF_mInt64(arrow_read_batch_size, "4096");
 CONF_mBool(brpc_socket_keepalive, "false");
 CONF_mBool(apply_del_vec_after_all_index_filter, "true");
 
+// python envs config
+// create time worker timeout
 CONF_mInt32(create_child_worker_timeout_ms, "1000");
+// config ENV PYTHONPATH
 CONF_Strings(python_envs, "");
+// report python worker STDERR to client
 CONF_Bool(report_python_worker_error, "true");
 CONF_Bool(python_worker_reuse, "true");
+CONF_Int32(python_worker_expire_time_sec, "300");
 
 } // namespace starrocks::config
