+++ be/src/udf/CMakeLists.txt
@@ -24,6 +24,7 @@ set(UDF_FILES
   java/java_udf.cpp
   java/java_data_converter.cpp
   java/java_native_method.cpp
+  python/env.cpp
   python/callstub.cpp
   udf_call_stub.cpp
 )
