+++ bin/stop_be.sh
@@ -38,6 +38,7 @@ export PID_DIR=`cd "$curdir"; pwd`
 source $STARROCKS_HOME/bin/common.sh
 
 export_env_from_conf $STARROCKS_HOME/conf/be.conf
+export_shared_envvars
 
 pidfile=$PID_DIR/be.pid
 
@@ -67,6 +68,16 @@ do
         ;;
     esac
 done
+
+# kill all python worker process
+find "${UDF_RUNTIME_DIR}" -maxdepth 1 -name 'pyworker*' -print0 | while IFS= read -r -d $'\0' worker; do
+    pid=$(echo "$worker" | sed -n 's/.*pyworker_\([0-9]*\).*/\1/p')
+    if [[ ! -z "$pid" ]]; then
+        kill -9 "$pid" > /dev/null
+        rm -- "$worker"
+    fi
+done
+
 
 if [ -f $pidfile ]; then
     pid=`cat $pidfile`
